+++ src/App.vue
@@ -4,8 +4,8 @@
 </template>
 
 <script>
-import MainNav from '@/components/MainNav.vue'
-import TheHero from '@/components/TheHero.vue'
+import MainNav from '@/components/Navigation/MainNav.vue'
+import TheHero from '@/components/JobSearch/TheHero.vue'
 
 export default {
   name: 'App',
