+++ tests/unit/components/Navigation/MainNav.test.js
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/vue'
 import userEvent from '@testing-library/user-event'
 
 // Component is the default export of the file
-import MainNav from '@/components/MainNav.vue'
+import MainNav from '@/components/Navigation/MainNav.vue'
 
 describe('MainNav', () => {
   const renderMainNav = () => {
